# swagger.json
{
  "basePath": "/api",
  "info": {
    "title": "chumbo-api",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
    "/leagues": {
      "get": {
        "summary": "Get all Chumbo seasons",
        "tags": [
          "league"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "league_id",
            "schema": {
              "type": "string"
            },
            "description": "Comma delimited list of league_ids"
          },
          {
            "in": "query",
            "name": "season",
            "schema": {
              "type": "string"
            },
            "description": "Get the league(s) for the given season(s)"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns individual league or array of leagues",
            "content": "application/json"
          }
        }
      }
    },
    "/league/{league_id}": {
      "get": {
        "summary": "Get single chumbo season",
        "tags": [
          "league"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "schema": {
              "type": "string",
              "example": "chumbo_2012"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns league object.",
            "content": "application/json"
          }
        }
      }
    },
    "/players": {
      "get": {
        "summary": "Get NFL players",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "player_id",
            "schema": {
              "type": "string"
            },
            "description": "Comma delimited list of player_ids"
          },
          {
            "in": "query",
            "name": "position",
            "schema": {
              "type": "string",
              "example": 50
            },
            "description": "Comma delimited list of player positions"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "description": "Limit response to this many players"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns array of players",
            "content": "application/json"
          }
        }
      }
    },
    "/players/{player_id}": {
      "get": {
        "summary": "Get NFL player by player_id",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "schema": {
              "type": "string",
              "example": 1387
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns player object.",
            "content": "application/json"
          }
        }
      }
    },
    "/teams": {
      "get": {
        "summary": "Get all Chumbo teams",
        "tags": [
          "team"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns every chumbo team (team)",
            "content": "application/json"
          }
        }
      }
    },
    "/league/{league_id}/teams": {
      "get": {
        "summary": "Get Chumbo teams for a league season",
        "tags": [
          "league",
          "team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "schema": {
              "type": "string",
              "example": "chumbo_2012"
            },
            "description": "League ID"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns all teams for a season",
            "content": "application/json"
          }
        }
      }
    },
    "/league/{league_id}/team/{roster_id}": {
      "get": {
        "summary": "Get information for a specific team in a season",
        "tags": [
          "league",
          "team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "schema": {
              "type": "string",
              "example": "chumbo_2012"
            }
          },
          {
            "in": "path",
            "name": "roster_id",
            "schema": {
              "type": "string",
              "example": 1
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns team object for given team.",
            "content": "application/json"
          }
        }
      }
    },
    "/owner/{owner_id}/teams": {
      "get": {
        "summary": "Get chumbo teams for a specific user",
        "tags": [
          "team",
          "owner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "owner_id",
            "schema": {
              "type": "string",
              "example": "chumbolegacy_euan"
            },
            "description": "The owner/user ID"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns teams for given user.",
            "content": "application/json"
          }
        }
      }
    },
    "/league/{league_id}/owners": {
      "get": {
        "summary": "Get Chumbo users for a league season",
        "tags": [
          "league",
          "owner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "schema": {
              "type": "string",
              "example": "chumbo_2012"
            },
            "description": "League ID"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns all users for a season",
            "content": "application/json"
          }
        }
      }
    },
    "/league/{league_id}/owner/{owner_id}": {
      "get": {
        "summary": "Get chumbo user for a specific season",
        "tags": [
          "league",
          "owner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "schema": {
              "type": "string",
              "example": "chumbo_2012"
            }
          },
          {
            "in": "path",
            "name": "owner_id",
            "schema": {
              "type": "string",
              "example": "chumbolegacy_jimmie"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns owner object for given user.",
            "content": "application/json"
          }
        }
      }
    },
    "/user/{user_id}": {
      "get": {
        "summary": "Get all chumbo seasons for a given user",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "schema": {
              "type": "string",
              "example": "chumbolegacy_euan"
            },
            "description": "The owner/user ID"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns users for given user.",
            "content": "application/json"
          }
        }
      }
    }
  },
  "definitions": {},
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}